Rewrite giantswarmClientPatcher.ts in TypeScript
